rename TransformerComponent to Transformer
