+++ lib/components/Animatable/animate.js
@@ -1,7 +1,7 @@
 const applyTransition = require('./dom/applyTransition')
 
 module.exports = function (options) {
-  // @TransformerComponent:animate(options)
+  // @Animatable:animate(options)
   //
   // Update CSS transition animation properties of the component.
   //
