+++ lib/components/Animatable/dom/applyTransition.js
@@ -1,5 +1,5 @@
 module.exports = function (elem, params) {
-  // @TransformerComponent.dom.applyTransition(el, params)
+  // @Transformer.dom.applyTransition(el, params)
   //
   // Update transition animation properties of an element.
   //
