+++ lib/components/BlockComponent/index.js
@@ -1,9 +1,9 @@
-const TransformerComponent = require('../TransformerComponent')
+const Transformer = require('../Transformer')
 
 const BlockComponent = function (element) {
   // @BlockComponent(element)
   //
-  // Inherits TransformerComponent
+  // Inherits Transformer
   //
   // BlockComponent is an abstract class for rectangular affine components that
   // cannot change their size or the size depends on the browser layout.
@@ -20,7 +20,7 @@ const BlockComponent = function (element) {
   //
 
   // Inherit
-  TransformerComponent.call(this, element)
+  Transformer.call(this, element)
 }
 
 const proto = BlockComponent.prototype
@@ -28,7 +28,7 @@ module.exports = BlockComponent
 proto.isBlockComponent = true
 
 // Inherit
-Object.assign(proto, TransformerComponent.prototype)
+Object.assign(proto, Transformer.prototype)
 
 // Methods
 proto.atBottomLeft = require('./atBottomLeft')
