+++ lib/components/Composite/index.js
@@ -1,9 +1,9 @@
-const TransformerComponent = require('../TransformerComponent')
+const Transformer = require('../Transformer')
 
 const Composite = function (element) {
   // @Composite(element)
   //
-  // Inherits TransformerComponent
+  // Inherits Transformer
   //
   // An abstract class for components that have no intrinsic size
   // but rather are a collection of child components and their descendants.
@@ -16,7 +16,7 @@ const Composite = function (element) {
   //
 
   // Use default anchor 0,0
-  TransformerComponent.call(this, element)
+  Transformer.call(this, element)
 
   // Add special class
   element.classList.add('affine-composite')
@@ -27,7 +27,7 @@ const proto = Composite.prototype
 proto.isComposite = true
 
 // Inherit
-Object.assign(proto, TransformerComponent.prototype)
+Object.assign(proto, Transformer.prototype)
 
 // Methods
 proto.getBoundingBox = require('./getBoundingBox')
