+++ lib/components/Edge/renderTransform.js
@@ -1,4 +1,4 @@
-const applyTransform2d = require('../TransformerComponent/dom/applyTransform2d')
+const applyTransform2d = require('../Transformer/dom/applyTransform2d')
 
 module.exports = function () {
   // @Edge:renderTransform()
