+++ lib/components/Hyperspace/renderTransform.js
@@ -1,4 +1,4 @@
-const applyTransform2d = require('../TransformerComponent/dom/applyTransform2d')
+const applyTransform2d = require('../Transformer/dom/applyTransform2d')
 const plane3 = require('affineplane').plane3
 
 module.exports = function () {
