+++ lib/components/InteractiveComponent/index.js
@@ -3,7 +3,7 @@ const InteractiveComponent = function () {
   //
   // Interaction methods for affine components.
   // Designed to be inherited by instance classes that
-  // also inherit TransformerComponent.
+  // also inherit Transformer.
   //
 
   // TODO Parameters:
