+++ lib/components/Transformer/atAnchor.js
@@ -2,7 +2,7 @@ const Point = require('../../geometry/Point')
 const point2 = require('affineplane').point2
 
 module.exports = function (alt) {
-  // @TransformerComponent:atAnchor(alt)
+  // @Transformer:atAnchor(alt)
   //
   // Get the transformer anchor point. This is practically the center of mass
   // of the component. Transformations that apply a pivot point, will default
