+++ lib/components/Transformer/dom/applyTransformOrigin.js
