+++ lib/components/Transformer/getDirection.js
@@ -1,7 +1,7 @@
 const Direction = require('../../geometry/Direction')
 
 module.exports = function (theta, phi) {
-  // @TransformerComponent:getDirection(theta[, phi])
+  // @Transformer:getDirection(theta[, phi])
   //
   // Get a direction from the spherical coordinate angles theta and phi
   // relative to the inner basis.
