+++ lib/components/Transformer/getDistanceTo.js
@@ -1,5 +1,5 @@
 module.exports = function (basis) {
-  // @TransformerComponent:getDistanceTo(basis)
+  // @Transformer:getDistanceTo(basis)
   //
   // Get distance between the transformer anchor to
   // the anchor of a Component.
