+++ lib/components/Transformer/getPosition.js
@@ -2,7 +2,7 @@ const point3 = require('affineplane').point3
 const Point = require('../../geometry/Point')
 
 module.exports = function () {
-  // @TransformerComponent:getPosition()
+  // @Transformer:getPosition()
   //
   // Get the position of the transformer anchor, represented on the parent.
   // The main difference between getPosition and atAnchor is that getPosition
