+++ lib/components/Transformer/getScale.js
@@ -1,7 +1,7 @@
 const Scale = require('../../geometry/Scale')
 
 module.exports = function () {
-  // @TransformerComponent:getScale()
+  // @Transformer:getScale()
   //
   // The scale of the basis.
   //
