+++ lib/components/Transformer/getVectorTo.js
@@ -1,5 +1,5 @@
 module.exports = function (basis) {
-  // @TransformerComponent:getVectorTo(basis)
+  // @Transformer:getVectorTo(basis)
   //
   // Get vector from this anchor to the anchor of the given basis.
   // If basis does not have an anchor, we assume (0,0) for the anchor.
