+++ lib/components/Transformer/index.js
@@ -1,7 +1,7 @@
 const Component = require('../Component')
 
-const TransformerComponent = function (element) {
-  // @TransformerComponent(element)
+const Transformer = function (element) {
+  // @Transformer(element)
   //
   // Inherits Component
   //
@@ -27,9 +27,9 @@ const TransformerComponent = function (element) {
   this.anchor = { x: 0, y: 0, z: 0 }
 }
 
-module.exports = TransformerComponent
-const proto = TransformerComponent.prototype
-proto.isTransformerComponent = true
+module.exports = Transformer
+const proto = Transformer.prototype
+proto.isTransformer = true
 
 // Inherit
 Object.assign(proto, Component.prototype)
@@ -47,7 +47,7 @@ proto.matchBasis = require('./matchBasis')
 proto.matchPoint = require('./matchPoint')
 proto.moveTo = require('./translateTo')
 proto.renderTransform = require('./renderTransform')
-// proto.resetTransformerComponent? moveTo with default params handles already.
+// proto.resetTransformer? moveTo with default params handles already.
 proto.rotateBy = require('./rotateBy')
 proto.rotateByDegrees = require('./rotateByDegrees')
 // TODO proto.rotateTo
