+++ lib/components/Transformer/match.js
@@ -3,7 +3,7 @@ const dir2 = fine.dir2
 const nudged = require('nudged')
 
 module.exports = function (params) {
-  // @TransformerComponent:match(params)
+  // @Transformer:match(params)
   //
   // Matching is a powerful way to position elements without the need to know
   // their exact rotation, scaling, or translation. Give one or more source
