+++ lib/components/Transformer/matchBasis.js
@@ -2,7 +2,7 @@ const fine = require('affineplane')
 const plane3 = fine.plane3
 
 module.exports = function (source, target) {
-  // @TransformerComponent:matchBasis(source, target)
+  // @Transformer:matchBasis(source, target)
   //
   // Transform the element so that source basis position on the element
   // will match the target basis.
