+++ lib/components/Transformer/matchPoint.js
@@ -1,5 +1,5 @@
 module.exports = function (source, target) {
-  // @TransformerComponent:matchPoint(source, target)
+  // @Transformer:matchPoint(source, target)
   //
   // Moves the element so that source point position on the element
   // matches the target position.
