+++ lib/components/Transformer/renderTransform.js
@@ -3,7 +3,7 @@ const counterTransformOrigin = require('./patch/counterTransformOrigin')
 const plane3 = require('affineplane').plane3
 
 module.exports = function (alt) {
-  // @TransformerComponent:renderTransform(alt)
+  // @Transformer:renderTransform(alt)
   //
   // Update the element.style.transform according to the basis placement.
   //
