+++ lib/components/Transformer/rotateBy.js
@@ -1,7 +1,7 @@
 const fine = require('affineplane')
 
 module.exports = function (radians, pivot) {
-  // @TransformerComponent:rotateBy(radians[, pivot])
+  // @Transformer:rotateBy(radians[, pivot])
   //
   // Rotate the basis around z axis.
   //
