+++ lib/components/Transformer/rotateByDegrees.js
@@ -1,7 +1,7 @@
 const factor = Math.PI / 180
 
 module.exports = function (degrees, pivot) {
-  // @TransformerComponent:rotateByDegrees(degrees[, pivot])
+  // @Transformer:rotateByDegrees(degrees[, pivot])
   //
   // Rotate the element by degrees around z axis of an optional pivot point.
   //
