+++ lib/components/Transformer/scaleBy.js
@@ -1,7 +1,7 @@
 const fine = require('affineplane')
 
 module.exports = function (multiplier, pivot) {
-  // @TransformerComponent:scaleBy(multiplier[, pivot])
+  // @Transformer:scaleBy(multiplier[, pivot])
   //
   // Scale the element.
   //
