+++ lib/components/Transformer/setAnchor.js
@@ -1,7 +1,7 @@
 const applyTransformOrigin = require('./dom/applyTransformOrigin')
 
 module.exports = function (point, y) {
-  // @TransformerComponent:setAnchor(point)
+  // @Transformer:setAnchor(point)
   //
   // Set the transformer anchor point. This does not move the element.
   //
