+++ lib/components/Transformer/setBasis.js
@@ -1,6 +1,6 @@
 module.exports = function (basis) {
-  // @TransformerComponent:setBasis(basis)
-  // @TransformerComponent:transformTo
+  // @Transformer:setBasis(basis)
+  // @Transformer:transformTo
   //
   // Transform this element so that its position, scale, and orientation match
   // the given basis.
