+++ lib/components/Transformer/setOrientation.js
@@ -1,5 +1,5 @@
 module.exports = function (orient, pivot) {
-  // @TransformerComponent:setOrientation(orient[, pivot])
+  // @Transformer:setOrientation(orient[, pivot])
   //
   // Rotate this basis so that its orientation matches the given orientation.
   // The rotation is performed around the given pivot point.
