+++ lib/components/Transformer/setScale.js
@@ -1,7 +1,7 @@
 const plane3 = require('affineplane').plane3
 
 module.exports = function (scale, origin) {
-  // @TransformerComponent:setScale(scale[, origin])
+  // @Transformer:setScale(scale[, origin])
   //
   // Dilate this basis so that its scale matches the given scale.
   // The dilation is performed about the given origin point.
