+++ lib/components/Transformer/snapGrid.js
@@ -1,5 +1,5 @@
 module.exports = function (grid, anchor) {
-  /// @TransformerComponent:snapGrid(grid, anchor)
+  /// @Transformer:snapGrid(grid, anchor)
   //
   // Snap the transformer anchor to the closest matching grid position.
   //
