+++ lib/components/Transformer/snapPixels.js
@@ -2,7 +2,7 @@ const plane3 = require('affineplane').plane3
 const abs = Math.abs
 
 module.exports = function (pivot) {
-  // @TransformerComponent:snapPixels([pivot])
+  // @Transformer:snapPixels([pivot])
   //
   // Coordinates of plane do not always match the pixel grid
   // of the screen device. This causes especially
