+++ lib/components/Transformer/transformBy.js
@@ -1,7 +1,7 @@
 const plane3 = require('affineplane').plane3
 
 module.exports = function (tr, origin) {
-  // @TransformerComponent:transformBy(tr[, origin])
+  // @Transformer:transformBy(tr[, origin])
   //
   // Transform (move) the basis in space.
   // For example, imagine a transform that rotates the basis
