+++ lib/components/Transformer/translateBy.js
@@ -3,7 +3,7 @@ const helm3 = fine.helm3
 const vec3 = fine.vec3
 
 module.exports = function (translation) {
-  // @TransformerComponent:translateBy(translation)
+  // @Transformer:translateBy(translation)
   //
   // Translate the element along x-, y-, and z-axis.
   // Translation does not rotate or scale the element.
