+++ lib/components/Transformer/translateTo.js
@@ -4,7 +4,7 @@ const vec3 = fine.vec3
 const plane3 = fine.plane3
 
 module.exports = function (point) {
-  // @TransformerComponent:translateTo(point)
+  // @Transformer:translateTo(point)
   //
   // Translate the element along x-, y-, and z-axis so that its anchor
   // matches the given point.
