+++ lib/components/Viewport/measureGroup.js
@@ -58,11 +58,11 @@ module.exports = function (group, filter) {
   let stackFilter
   if (typeof filter === 'function') {
     stackFilter = (stackItem) => {
-      return stackItem.node.isTransformerComponent && filter(stackItem.node)
+      return stackItem.node.isTransformer && filter(stackItem.node)
     }
   } else {
     stackFilter = (stackItem) => {
-      return stackItem.node.isTransformerComponent
+      return stackItem.node.isTransformer
     }
   }
