+++ lib/components/Viewport/measureOne.js
@@ -12,7 +12,7 @@ module.exports = function (component) {
   //
   // Parameters:
   //   component
-  //     a TransformerComponent
+  //     a Transformer
   //
   // Returns:
   //   a Measurement
