+++ lib/components/Viewport/renderTransform.js
@@ -9,7 +9,7 @@ module.exports = function (alt) {
   // Parameters:
   //   alt
   //     optional plane2 transition to render instead of this.tran.
-  //     .. See TransformerComponent:renderTransform for details.
+  //     .. See Transformer:renderTransform for details.
   //
 
   // Delegate to each space.
