+++ lib/components/index.js
@@ -17,12 +17,12 @@
 //
 // **Abstract components:**
 // - Component is a coordinate system in affine space.
-// - TransformerComponent is a Component that is transformable.
-// - BlockComponent is a TransformerComponent that has rectangular
+// - Transformer is a Component that is transformable.
+// - BlockComponent is a Transformer that has rectangular
 //   .. boundaries and size.
 // - FrameComponent is a BlockComponent that has a changeable size.
 // - ControlComponent is a FrameComponent that stays fixed to the viewport.
-// - InteractiveComponent is a TransformerComponent that can capture
+// - InteractiveComponent is a Transformer that can capture
 //   .. touch and mouse gestures.
 //
 // **Inheritance chart:**
@@ -38,7 +38,7 @@ exports.Composite = require('./Composite')
 exports.ControlComponent = require('./ControlComponent')
 exports.FrameComponent = require('./FrameComponent')
 exports.InteractiveComponent = require('./InteractiveComponent')
-exports.TransformerComponent = require('./TransformerComponent')
+exports.Transformer = require('./Transformer')
 
 // Instance classes
 exports.Arc = require('./Arc')
