+++ lib/geometry/Basis/getMatchedOuter.js
@@ -9,7 +9,7 @@ module.exports = function (targetBasis) {
   // and you know a basis relative to that element and you also know
   // where that basis should be space. Then the resulting basis is the basis
   // to which you should move the element.
-  // See also TransformerComponent:matchBasis
+  // See also Transformer:matchBasis
   //
   // Parameters:
   //   targetBasis
