+++ lib/interaction/Hold/index.js
@@ -22,7 +22,7 @@ const HoldGesture = function (source, target, options) {
   //     an InteractiveComponent. The hold input source.
   //     .. The source begins to emit hold events.
   //   target
-  //     a TransformerComponent. The target for the hold effects, if any.
+  //     a Transformer. The target for the hold effects, if any.
   //   options
   //     holdDuration
   //       optional number, default 500.
