+++ lib/interaction/KeyboardPan/index.js
@@ -9,7 +9,7 @@ const KeyboardPan = function (source, target, options) {
   //   source
   //     an InteractiveComponent. The keyboard input source.
   //   target
-  //     a TransformerComponent. The target to apply the panning.
+  //     a Transformer. The target to apply the panning.
   //   options
   //     optional object with properties:
   //       step
