+++ lib/interaction/KeyboardZoom/index.js
@@ -9,7 +9,7 @@ const KeyboardZoom = function (source, target, options) {
   //   source
   //     an InteractiveComponent. The keyboard input source.
   //   target
-  //     a TransformerComponent. The target to zoom towards.
+  //     a Transformer. The target to zoom towards.
   //   options
   //     optional object with properties:
   //       step
