+++ lib/interaction/Pinch/index.js
@@ -10,7 +10,7 @@ const PinchGesture = function (source, target, options) {
   //     an InteractiveComponent. Get gesture input from this component.
   //     .. The source begins to emit pinch events.
   //   target
-  //     a TransformerComponent.
+  //     a Transformer.
   //     .. Apply gesture effects to this component.
   //   options, object with properties:
   //     freedom
