+++ lib/interaction/Slide/index.js
@@ -9,7 +9,7 @@ const SlideGesture = function (source, target, options) {
   //     an InteractiveComponent. The slide input source.
   //     .. The source will emit slide events.
   //   target
-  //     a TransformerComponent.
+  //     a Transformer.
   //     .. The target for the slide transformation and effects.
   //   options
   //     direction
