+++ lib/interaction/Tap/index.js
@@ -8,7 +8,7 @@ const TapGesture = function (source, target, options) {
   //     an InteractiveComponent. The tap input source.
   //     .. The source will emit 'tap' events.
   //   target
-  //     a TransformerComponent. The target for the optional tap effect.
+  //     a Transformer. The target for the optional tap effect.
   //   options
   //     effect
   //       TODO string, one of 'shrink', 'shake', 'down'.
